Reorganize folder structure and abtract app screen logic
